requirements.txt
python-dotenv~=0.21.0
#pyodbc==4.0.34
django-mssql-backend==2.8.1

Django~=3.2
django-phonenumber-field~=5.1.0
django-mptt~=0.12.0
django-extensions~=3.1.3
django-json-widget~=1.1.1
django-crispy-forms~=1.11.2
django-admin-sortable2==1.0
django-redis==5.2.0
django-sequences==2.6
django-solo==2.0.0
django-webpack-loader==1.6.0
django-log-request-id==2.0.0
Pillow==10.0.0
mssql-django==1.3
xhtml2pdf==0.2.11
python-docx==0.8.11
clr==1.0.3
num2words==0.5.12
bcrypt==4.0.1
translate~=3.6.1
unoconv==0.9.0
pycryptodome==3.18.0
cryptography==41.0.2

# Django REST Framework
djangorestframework~=3.12.4
djangorestframework-jwt==1.11.0
drf-spectacular[sidecar]
django-cors-headers~=3.7.0
djangorestframework-simplejwt
drf-writable-nested~=0.6.2
drf-extra-fields~=3.4.0
django-filter~=2.4.0
markdown~=3.4.1
djoser~=2.1.0
django-rest-swagger~=2.2.0
django-seed==0.3.1

# Celery
celery==5.3.0
django-celery-beat==2.5.0

# Utils
phonenumbers~=8.12.21
pyotp~=2.6.0
psycopg2-binary~=2.9.2
daphne==3.0.1
gunicorn==20.1.0
django-environ==0.9.0
channels==3.0.4
channels-redis==3.3.1
django-redis==5.2.0

requests~=2.31.0
docx~=0.2.4
reportlab~=3.6.13
httpx==0.28.1
